+++ alugafacil/src/main/java/com/example/alugafacil/alugafacil/repositories/InquilinoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InquilinoRepository extends JpaRepository<Inquilino, Integer> {
+public interface InquilinoRepository extends JpaRepository<Inquilino, Long> {
 }
